+++ pydfc/dfc_methods/sliding_window_clustr.py
@@ -228,7 +228,7 @@ class SLIDING_WINDOW_CLUSTR(BaseDFCMethod):
             n_clusters=self.params["n_states"],
             n_regions=dFC_raw.n_regions,
         )
-        self.Z = self.kmeans_.predict(self.dFC_mat2vec(SW_dFC))
+        self.Z = self.kmeans_.predict(self.dFC_mat2vec(SW_dFC).astype(np.float32))
 
         # mean activation of states
         self.set_mean_activity(time_series)
@@ -270,7 +270,7 @@ class SLIDING_WINDOW_CLUSTR(BaseDFCMethod):
             # Z = self.clusters_lst2idx(self.kmeans_.get_clusters())
         else:
             ########### Euclidean Clustering ##############
-            Z = self.kmeans_.predict(F)
+            Z = self.kmeans_.predict(F.astype(np.float32))
 
         # record time
         self.set_dFC_assess_time(time.time() - tic)
